+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 6.2v4
 - Fixed incorrect mission threat counter information when threats were encountered multiple times
 - Fixed shipyard bonus effect areas not showing on non-English locales
-- Fixed error when double clicking on follower list when not currently building a party.
+- Fixed error when double clicking on follower list when not currently building a party
+- Updated follower learnable counters mappings
 
 # 6.2v3
 - Fixed lua error when follower list was too short