+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fixed incorrect mission threat counter information when threats were encountered multiple times
 - Fixed shipyard bonus effect areas not showing on non-English locales
 - Fixed error when double clicking on follower list when not currently building a party
+- Fixed error when clicking threat buttons on recruit selection
 - Updated follower learnable counters mappings
 
 # 6.2v3