+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # 6.2v4
 - New: Reward counts in mission tooltips now consider follower bonuses to currency and gold yields
+- New: Fixed Blizzard bug causing shipyard bonus effect areas not showing on non-English locales
 - Fixed incorrect mission threat counter information when threats were encountered multiple times
 - Fixed shipyard bonus effect areas not showing on non-English locales
 - Fixed error when double clicking on follower list when not currently building a party
 - Fixed error when clicking threat buttons on recruit selection
 - Updated follower learnable counters mappings
+- Reward counts in mission tooltips now consider follower bonuses to currency and gold yields
 
 # 6.2v3
 - Fixed lua error when follower list was too short